# packages/api/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "SelfFeed API",
		"version": "1.0.0",
		"description": "Stable HTTP API for the SelfFeed web client and future mobile clients."
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		},
		"schemas": {
			"ApiError": {
				"type": "object",
				"required": [
					"error"
				],
				"properties": {
					"error": {
						"type": "object",
						"required": [
							"code",
							"message"
						],
						"properties": {
							"code": {
								"type": "string"
							},
							"message": {
								"type": "string"
							},
							"details": {}
						}
					}
				}
			},
			"User": {
				"type": "object",
				"required": [
					"id",
					"email",
					"role",
					"isActive",
					"createdAt",
					"updatedAt"
				],
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"role": {
						"type": "string",
						"enum": [
							"admin",
							"user"
						]
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"AccessToken": {
				"type": "object",
				"required": [
					"accessToken"
				],
				"properties": {
					"accessToken": {
						"type": "string"
					}
				}
			},
			"AuthResponse": {
				"type": "object",
				"required": [
					"user",
					"tokens"
				],
				"properties": {
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"tokens": {
						"$ref": "#/components/schemas/AccessToken"
					}
				}
			},
			"RefreshResponse": {
				"type": "object",
				"required": [
					"tokens"
				],
				"properties": {
					"tokens": {
						"$ref": "#/components/schemas/AccessToken"
					}
				}
			},
			"AuthSession": {
				"type": "object",
				"required": [
					"id",
					"deviceName",
					"createdAt",
					"lastSeenAt",
					"current"
				],
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"deviceName": {
						"type": "string"
					},
					"clientId": {
						"type": [
							"string",
							"null"
						]
					},
					"ipAddress": {
						"type": [
							"string",
							"null"
						]
					},
					"userAgent": {
						"type": [
							"string",
							"null"
						]
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"lastSeenAt": {
						"type": "string",
						"format": "date-time"
					},
					"current": {
						"type": "boolean"
					}
				}
			},
			"AuthSessionsResponse": {
				"type": "object",
				"required": [
					"sessions"
				],
				"properties": {
					"sessions": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/AuthSession"
						}
					}
				}
			},
			"AdminUsersResponse": {
				"type": "object",
				"required": [
					"users",
					"cursor",
					"hasMore"
				],
				"properties": {
					"users": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/User"
						}
					},
					"cursor": {
						"type": [
							"string",
							"null"
						]
					},
					"hasMore": {
						"type": "boolean"
					}
				}
			},
...
			"AppSettings": {
				"type": "object",
				"required": [
					"registrationLocked"
				],
				"properties": {
					"registrationLocked": {
						"type": "boolean"
					}
				}
			},
...
		}
	},
	"paths": {
...
		"/auth/register": {
			"post": {
				"tags": [
					"Auth"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"email",
									"password"
								],
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									},
									"password": {
										"type": "string",
										"minLength": 8
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/AuthResponse"
										}
									}
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"tags": [
					"Auth"
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"email",
									"password"
								],
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									},
									"password": {
										"type": "string"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/AuthResponse"
										}
									}
								}
							}
						}
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
...
		"/auth/refresh": {
			"post": {
				"tags": [
					"Auth"
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/RefreshResponse"
										}
									}
								}
							}
						}
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/auth/me": {
			"get": {
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/User"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/auth/sessions": {
			"get": {
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/AuthSessionsResponse"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/auth/sessions/{sessionId}": {
			"delete": {
				"tags": [
					"Auth"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "sessionId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {
											"type": "object",
											"properties": {
												"success": {
													"type": "boolean"
												}
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiError"
								}
							}
						}
					}
				}
			}
		},
		"/admin/settings": {
			"get": {
				"tags": [
					"Admin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/AppSettings"
										}
									}
								}
							}
						}
					}
				}
			},
			"patch": {
				"tags": [
					"Admin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AppSettings"
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/AppSettings"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/admin/users": {
			"get": {
				"tags": [
					"Admin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 100,
							"default": 25
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"schema": {
							"type": "string",
							"pattern": "^\\d+$"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/AdminUsersResponse"
										}
									}
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"Admin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"email",
									"password"
								],
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									},
									"password": {
										"type": "string",
										"minLength": 8
									},
									"role": {
										"type": "string",
										"enum": [
											"admin",
											"user"
										]
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/User"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/admin/users/{userId}": {
			"patch": {
				"tags": [
					"Admin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"minProperties": 1,
								"properties": {
									"role": {
										"type": "string",
										"enum": [
											"admin",
											"user"
										]
									},
									"isActive": {
										"type": "boolean"
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/User"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/admin/users/{userId}/reset-password": {
			"post": {
				"tags": [
					"Admin"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"required": [
									"password"
								],
								"properties": {
									"password": {
										"type": "string",
										"minLength": 8,
										"maxLength": 128
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"required": [
										"data"
									],
									"properties": {
										"data": {
											"$ref": "#/components/schemas/User"
										}
									}
								}
							}
						}
					}
				}
			}
		},
...
	}
}